+++ tests/ContainerBuilderTest.php
@@ -7,10 +7,11 @@ namespace Phetit\DependencyInjection\Tests;
 use Phetit\DependencyInjection\ContainerBuilder;
 use Phetit\DependencyInjection\Tests\Fixtures\Service;
 use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
 
 class ContainerBuilderTest extends TestCase
 {
-    public function testRegisterMethod(): void
+    public function testRegisterMethodShouldAddServiceResolver(): void
     {
         $builder = new ContainerBuilder();
 
@@ -25,7 +26,7 @@ class ContainerBuilderTest extends TestCase
         self::assertSame($serviceOne, $serviceTwo);
     }
 
-    public function testFactoryMethod(): void
+    public function testFactoryMethodShouldAddFactoryServiceResolver(): void
     {
         $builder = new ContainerBuilder();
 
@@ -38,5 +39,15 @@ class ContainerBuilderTest extends TestCase
         self::assertInstanceOf(Service::class, $serviceTwo);
 
         self::assertNotSame($serviceOne, $serviceTwo);
+    }
+
+    public function testShouldInjectAutoReference(): void
+    {
+        $builder = new ContainerBuilder();
+
+        $container = $builder->get(ContainerInterface::class);
+
+        self::assertInstanceOf(ContainerBuilder::class, $container);
+        self::assertSame($builder, $container);
     }
 }